pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = "Data engineer Alkemy Challenger"
authors = ["MorillaGit <cristian.morilla@mi.unc.edu.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^1.4.43"
pandas = "^1.5.1"
psycopg2 = "^2.9.5"
psycopg2-binary = "^2.9.5"
python-decouple = "^3.6"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
